Improve ch-dropdown's component structure
