+++ src/components/dropdown/internal/dropdown.scss
@@ -1,4 +1,4 @@
-@import "../../common/_base";
+@import "../../../common/_base";
 
 $z-index-gx-navbar-item: 107; // Same as $z-index-gx-navbar-item from w-c-l
