+++ src/components/dropdown/internal/types.ts
@@ -1,0 +1,24 @@
+export type DropdownAlign =
+  | "OutsideStart"
+  | "InsideStart"
+  | "Center"
+  | "InsideEnd"
+  | "OutsideEnd";
+
+export type DropdownPosition =
+  | "OutsideStart_OutsideStart"
+  | "InsideStart_OutsideStart"
+  | "Center_OutsideStart"
+  | "InsideEnd_OutsideStart"
+  | "OutsideEnd_OutsideStart"
+  | "OutsideStart_InsideStart"
+  | "OutsideEnd_InsideStart"
+  | "OutsideStart_Center"
+  | "OutsideEnd_Center"
+  | "OutsideStart_InsideEnd"
+  | "OutsideEnd_InsideEnd"
+  | "OutsideStart_OutsideEnd"
+  | "InsideStart_OutsideEnd"
+  | "Center_OutsideEnd"
+  | "InsideEnd_OutsideEnd"
+  | "OutsideEnd_OutsideEnd";
